pyproject.toml
[project]
name = "agetic-conversation-bot"
version = "0.1.0"
description = "Agetic Conversation Bot with TTS and STT and LLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "elevenlabs>=2.15.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "noisereduce>=3.0.3",
    "numpy>=2.2.6",
    "openai>=1.107.0",
    "openai-whisper>=20250625",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "simpleaudio>=1.0.4",
    "speechrecognition>=3.14.3",
]
